fix back

+++ server/src/restaurant/restaurant.service.ts
@@ -1,4 +1,4 @@
-import {Injectable} from '@nestjs/common';
+import {Injectable, NotFoundException} from '@nestjs/common';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 
@@ -6,6 +6,7 @@ import {InjectRepository} from "@nestjs/typeorm";
 import {Repository} from "typeorm";
 import {Restaurant} from "./entities/restaurant.entity";
 import {Menu} from "../menu/entities/menu.entity";
+import {Vote} from "../vote/entities/vote.entity";
 
 
 
@@ -19,19 +20,30 @@ export class RestaurantService {
     return  this.RestaurantRepository.save(createRestaurantDto);
   }
 
-  findAll() {
-    return  this.RestaurantRepository.find();
-  }
-
+    async findAll(): Promise<Restaurant[]> {
+        return await this.RestaurantRepository.find({
+                relations : {
+                    Menus: true,
+                    Students:true,
+                    Votes:true
+                }
+            },
+        );
+    }
 
   findOne(id: string): Promise<Restaurant> {
     return this.RestaurantRepository.findOne({
       where: {
         id: id
+      },
+      relations : {
+            Menus: true,
+            Students:true,
+            Votes:true
       }
     }).then(Restaurant => {
       if(!Restaurant){
-        throw new Error("Restaurant not found");
+        throw new NotFoundException("Restaurant not found");
       }
       return Restaurant;
     });
@@ -39,10 +51,21 @@ export class RestaurantService {
   }
 
   update(id: string, updateRestaurantDto: UpdateRestaurantDto) {
+      //chek if the restaurant exists
+        const restaurant = this.RestaurantRepository.findOne({where : {id : id}});
+        if(!restaurant){
+            throw new NotFoundException("Restaurant not found");
+        }
+
     return this.RestaurantRepository.update(id, updateRestaurantDto);
   }
 
   remove(id: string) {
+
+    const restaurant = this.RestaurantRepository.findOne({where : {id : id}});
+    if(!restaurant){
+      throw new NotFoundException("Restaurant not found");
+    }
     return this.RestaurantRepository.delete(id);
   }
 }
